web_crawler/requirements.txt
pandas==2.2.3
PyYAML==6.0.2
selenium==4.28.1
tqdm==4.67.1
